move renderElementToSvg to export
